Isolate Dom from the logic
