+++ src/createTask.js
@@ -1,4 +1,4 @@
-class Task {
+ export default class Task {
   constructor(title, description, dueDate, priority) {
     this.title = title;
     this.description = description;
@@ -7,21 +7,5 @@ class Task {
   }
 }
 
-export function clearInputs() {
-  document.querySelector('#task-title').value = '';
-  document.querySelector('#task-description').value = '';
-  document.querySelector('#dueDate').value = '';
-  document.querySelector('#priority').value = '';
-}
 
-export function newTask() {
-  const title = document.querySelector('#task-title').value;
-  const description = document.querySelector('#task-description').value;
-  const dueDate = document.querySelector('#dueDate').value;
-  const priority = document.querySelector('#priority').value;
-  const newT = new Task(title, description, dueDate, priority);
-  clearInputs();
-  return newT;
-}
 
-export default newTask;
